+++ indico/modules/events/registration/controllers/display.py
@@ -9,7 +9,7 @@ from uuid import UUID
 
 from flask import flash, jsonify, redirect, request, session
 from sqlalchemy.orm import contains_eager, joinedload, lazyload, load_only, subqueryload
-from werkzeug.exceptions import BadRequest, Forbidden, NotFound, UnprocessableEntity
+from werkzeug.exceptions import BadRequest, Forbidden, NotFound
 
 from indico.core.db import db
 from indico.modules.auth.util import redirect_to_login
@@ -18,7 +18,9 @@ from indico.modules.events.controllers.base import RegistrationRequired, RHDispl
 from indico.modules.events.models.events import EventType
 from indico.modules.events.payment import payment_event_settings
 from indico.modules.events.registration import registration_settings
-from indico.modules.events.registration.controllers import CheckEmailMixin, RegistrationEditMixin, RegistrationFormMixin
+from indico.modules.events.registration.controllers import (CheckEmailMixin, RegistrationEditMixin,
+                                                            RegistrationFormMixin, UploadRegistrationFileMixin,
+                                                            UploadRegistrationPictureMixin)
 from indico.modules.events.registration.models.form_fields import RegistrationFormFieldData, RegistrationFormItem
 from indico.modules.events.registration.models.forms import RegistrationForm
 from indico.modules.events.registration.models.invitations import InvitationState, RegistrationInvitation
@@ -27,12 +29,10 @@ from indico.modules.events.registration.models.registrations import Registration
 from indico.modules.events.registration.notifications import notify_registration_state_update
 from indico.modules.events.registration.util import (create_registration, generate_ticket,
                                                      get_event_regforms_registrations, get_flat_section_submission_data,
-                                                     get_initial_form_values, get_user_data, make_registration_schema,
-                                                     process_registration_picture)
+                                                     get_initial_form_values, get_user_data, make_registration_schema)
 from indico.modules.events.registration.views import (WPDisplayRegistrationFormConference,
                                                       WPDisplayRegistrationFormSimpleEvent,
                                                       WPDisplayRegistrationParticipantList)
-from indico.modules.files.controllers import UploadFileMixin
 from indico.modules.receipts.models.files import ReceiptFile
 from indico.modules.users.util import send_avatar, send_default_avatar
 from indico.util.fs import secure_filename
@@ -401,14 +401,8 @@ class RHRegistrationForm(InvitationMixin, RHRegistrationFormRegistrationBase):
                                                captcha_settings=get_captcha_settings())
 
 
-class RHUploadRegistrationFile(UploadFileMixin, InvitationMixin, RHRegistrationFormBase):
-    """
-    Upload a file from a registration form.
-
-    Regform file fields do not wait for the regform to be submitted,
-    but upload the selected files immediately, saving just the generated uuid.
-    Only this uuid is then sent when the regform is submitted.
-    """
+class RHUploadRegistrationFile(UploadRegistrationFileMixin, InvitationMixin, RHRegistrationFormBase):
+    """Upload a file from a registration form."""
 
     ALLOW_PROTECTED_EVENT = True
 
@@ -431,20 +425,9 @@ class RHUploadRegistrationFile(UploadFileMixin, InvitationMixin, RHRegistrationF
                 if not self.invitation or not self.invitation.skip_access_check:
                     raise
 
-    def get_file_context(self):
-        return 'event', self.event.id, 'regform', self.regform.id, 'registration'
 
-
-class RHUploadRegistrationPicture(RHUploadRegistrationFile):
+class RHUploadRegistrationPicture(UploadRegistrationPictureMixin, RHUploadRegistrationFile):
     """Upload a picture from a registration form."""
-
-    def _save_file(self, file, stream):
-        if not (resized_image_stream := process_registration_picture(stream)):
-            raise UnprocessableEntity('Could not process image, it may be corrupted or too big')
-        return super()._save_file(file, resized_image_stream)
-
-    def get_file_metadata(self):
-        return {'registration_picture_checked': True}
 
 
 class RHRegistrationDisplayEdit(RegistrationEditMixin, RHRegistrationFormRegistrationBase):